[refactor] 도메인 중심의 디렉토리 구조로 변경
