+++ src/main/java/com/example/spring/domain/dateSpot/enums/Category.java
@@ -1,4 +1,4 @@
-package com.example.spring.domain.enums;
+package com.example.spring.domain.dateSpot.enums;
 
 public enum Category {
     RESTAURANT,
