+++ src/main/java/com/example/spring/domain/dateSpot/enums/InOut.java
@@ -1,0 +1,5 @@
+package com.example.spring.domain.dateSpot.enums;
+
+public enum InOut {
+    IN, OUT
+}
