--- src/main/java/com/example/spring/domain/enums/Gender.java
@@ -1,5 +1,0 @@
-package com.example.spring.domain.enums;
-
-public enum Gender {
-    MALE, FEMALE
-}
