--- src/main/java/com/example/spring/domain/enums/Status.java
@@ -1,5 +1,0 @@
-package com.example.spring.domain.enums;
-
-public enum Status {
-    ACTIVE, INACTIVE
-}
