+++ src/main/java/com/example/spring/domain/member/domain/Qna.java
@@ -1,7 +1,7 @@
-package com.example.spring.domain;
+package com.example.spring.domain.member.domain;
 
-import com.example.spring.domain.common.BaseEntity;
-import com.example.spring.domain.enums.AnswerStatus;
+import com.example.spring.global.baseEntity.BaseEntity;
+import com.example.spring.domain.member.enums.AnswerStatus;
 import jakarta.persistence.*;
 import lombok.*;
