+++ src/main/java/com/example/spring/domain/member/enums/PlusMinus.java
@@ -1,0 +1,5 @@
+package com.example.spring.domain.member.enums;
+
+public enum PlusMinus {
+    PLUS, MINUS
+}
