+++ src/main/java/com/example/spring/domain/member/enums/Status.java
@@ -1,0 +1,5 @@
+package com.example.spring.domain.member.enums;
+
+public enum Status {
+    ACTIVE, INACTIVE
+}
